+++ src/main/kotlin/com/pestphp/pest/PestIcons.kt
@@ -9,6 +9,7 @@ object PestIcons {
     val RUN_SINGLE_TEST = IconLoader.getIcon("/run@${getThemeString()}.svg", PestIcons.javaClass)
     val CONFIG = IconLoader.getIcon("/config@${getThemeString()}.svg", PestIcons.javaClass)
     val FILE = IconLoader.getIcon("/file@${getThemeString()}.svg", PestIcons.javaClass)
+    val DATASET_FILE = IconLoader.getIcon("/dataset@${getThemeString()}.svg", PestIcons.javaClass)
 
     private fun isDarkMode() = LafManager.getInstance().currentLookAndFeel == DarculaLookAndFeelInfo()
     private fun getThemeString(): String {