+++ src/test/resources/com/pestphp/pest/Dataset.php
@@ -1,0 +1,6 @@
+<?php
+
+dataset('emails', [
+    'enunomaduro@gmail.com',
+    'oliver@nybroe.dev'
+]);