+++ src/test/resources/com/pestphp/pest/TestWithDataset.php
@@ -1,0 +1,10 @@
+<?php
+
+dataset('emails', [
+    'enunomaduro@gmail.com',
+    'oliver@nybroe.dev'
+]);
+
+it('basic', function () {
+    $this->assertTrue(true);
+});